+++ packages/client/src/base/default.module.ts
@@ -22,6 +22,7 @@ import {
     MoveCommand,
     SetDirtyStateAction,
     SetEditModeAction,
+    SetModelCommand,
     TYPES,
     bindAsService,
     bindOrRebind,
@@ -31,10 +32,12 @@ import {
 } from '@eclipse-glsp/sprotty';
 import '@vscode/codicons/dist/codicon.css';
 import '../../css/glsp-sprotty.css';
+import { bindContributionProvider } from '../utils/contribution-provider';
 import { GLSPActionDispatcher } from './action-dispatcher';
 import { GLSPActionHandlerRegistry } from './action-handler-registry';
 import { GLSPCommandStack } from './command-stack';
 import { EditorContextService } from './editor-context-service';
+import { FeedbackAwareSetModelCommand } from './feedback';
 import { ModifyCssFeedbackCommand } from './feedback/css-feedback';
 import { FeedbackActionDispatcher } from './feedback/feedback-action-dispatcher';
 import { FeedbackAwareUpdateModelCommand } from './feedback/update-model-command';
@@ -73,6 +76,7 @@ export const defaultModule = new FeatureModule((bind, unbind, isBound, rebind, .
     // Model update initialization ------------------------------------
     bind(TYPES.IFeedbackActionDispatcher).to(FeedbackActionDispatcher).inSingletonScope();
     configureCommand(context, FeedbackAwareUpdateModelCommand);
+    rebind(SetModelCommand).to(FeedbackAwareSetModelCommand);
 
     bind(GLSPMouseTool).toSelf().inSingletonScope();
     bindOrRebind(context, MouseTool).toService(GLSPMouseTool);
@@ -84,6 +88,9 @@ export const defaultModule = new FeatureModule((bind, unbind, isBound, rebind, .
     bindOrRebind(context, TYPES.ICommandStack).to(GLSPCommandStack).inSingletonScope();
     bind(GLSPActionDispatcher).toSelf().inSingletonScope();
     bindOrRebind(context, TYPES.IActionDispatcher).toService(GLSPActionDispatcher);
+
+    bindContributionProvider(bind, TYPES.ActionHandlerRegistration);
+    bindContributionProvider(bind, TYPES.IActionHandlerInitializer);
     bind(GLSPActionHandlerRegistry).toSelf().inSingletonScope();
     bindOrRebind(context, ActionHandlerRegistry).toService(GLSPActionHandlerRegistry);