+++ packages/client/src/base/feedback/update-model-command.ts
@@ -13,25 +13,20 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
-import { inject, injectable, optional } from 'inversify';
 import {
-    ActionHandlerRegistry,
     Animation,
-    Command,
-    CommandActionHandler,
     CommandExecutionContext,
     CommandReturn,
+    GModelRoot,
     ILogger,
     MorphEdgesAnimation,
-    GModelRoot,
     TYPES,
     UpdateAnimationData,
     UpdateModelAction,
-    UpdateModelCommand,
-    toTypeGuard
+    UpdateModelCommand
 } from '@eclipse-glsp/sprotty';
+import { inject, injectable, optional } from 'inversify';
 import { IFeedbackActionDispatcher } from './feedback-action-dispatcher';
-import { FeedbackCommand } from './feedback-command';
 
 /**
  * A special {@link UpdateModelCommand} that retrieves all registered {@link Action}s from the {@link IFeedbackActionDispatcher}
@@ -47,48 +42,16 @@ export class FeedbackAwareUpdateModelCommand extends UpdateModelCommand {
     @optional()
     protected feedbackActionDispatcher: IFeedbackActionDispatcher;
 
-    protected actionHandlerRegistry?: ActionHandlerRegistry;
-
-    constructor(
-        @inject(TYPES.Action) action: UpdateModelAction,
-        @inject(TYPES.ActionHandlerRegistryProvider)
-        actionHandlerRegistryProvider: () => Promise<ActionHandlerRegistry>
-    ) {
+    constructor(@inject(TYPES.Action) action: UpdateModelAction) {
         super({ animate: true, ...action });
-        actionHandlerRegistryProvider().then(registry => (this.actionHandlerRegistry = registry));
     }
 
     protected override performUpdate(oldRoot: GModelRoot, newRoot: GModelRoot, context: CommandExecutionContext): CommandReturn {
-        if (this.feedbackActionDispatcher && this.actionHandlerRegistry) {
-            // Create a temporary context which defines the `newRoot` as `root`
-            // This way we do not corrupt the redo/undo behavior of the super class
-            const tempContext: CommandExecutionContext = {
-                ...context,
-                root: newRoot
-            };
-
-            const feedbackCommands = this.getFeedbackCommands(this.actionHandlerRegistry);
-            feedbackCommands.forEach(command => command.execute(tempContext));
-        }
-
+        // Create a temporary context which defines the `newRoot` as `root`
+        // This way we do not corrupt the redo/undo behavior of the super class
+        const tempContext: CommandExecutionContext = { ...context, root: newRoot };
+        this.feedbackActionDispatcher?.applyFeedbackCommands(tempContext);
         return super.performUpdate(oldRoot, newRoot, context);
-    }
-
-    protected getFeedbackCommands(registry: ActionHandlerRegistry): Command[] {
-        const result: Command[] = [];
-        this.feedbackActionDispatcher.getRegisteredFeedback().forEach(action => {
-            const commands = registry
-                .get(action.kind)
-                .filter<CommandActionHandler>(toTypeGuard(CommandActionHandler))
-                .map(handler => handler.handle(action));
-            result.push(...commands);
-        });
-        // sort commands descanting by priority
-        return result.sort((a, b) => this.getPriority(b) - this.getPriority(a));
-    }
-
-    protected getPriority(command: Partial<FeedbackCommand>): number {
-        return command.priority ?? 0;
     }
 
     // Override the `createAnimations` implementation and remove the animation for edge morphing. Otherwise routing & reconnect