+++ packages/glsp-sprotty/src/types.ts
@@ -14,6 +14,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
 import { TYPES as SprottyTYPES } from 'sprotty';
+
 /**
  * Reexport of the TYPES namespace of sprotty augments with additional GLSP specific service
  * identifiers.
@@ -46,5 +47,6 @@ export const TYPES = {
     ILocalElementNavigator: Symbol('ILocalElementNavigator'),
     IDiagramOptions: Symbol('IDiagramOptions'),
     IDiagramStartup: Symbol('IDiagramStartup'),
-    IToolManager: Symbol('IToolManager')
+    IToolManager: Symbol('IToolManager'),
+    IContributionProvider: Symbol('IContributionProvider')
 };